requirements.txt
Flask==0.10
Flask-OAuthlib>=0.5.0
PyYAML==3.11
nose==1.3.7
gcloud==0.18.1
PyJWT==1.4.2